Clean up multiple worker loading
