+++ supirfactor_dynamical/tests/test_loader.py
@@ -3,11 +3,8 @@ import anndata as ad
 import numpy as np
 import numpy.testing as npt
 import pandas as pd
-import pandas.testing as pdt
 import scipy.sparse as sps
 import torch
-import tempfile
-import os
 
 from torch.utils.data import DataLoader
 
@@ -17,16 +14,13 @@ from ._stubs import (
 )
 
 from supirfactor_dynamical.datasets import (
-    TimeDataset,
-    H5ADDataset,
-    H5ADDatasetIterable,
-    H5ADDatasetStratified,
-    H5ADDatasetObsStratified,
-    StackIterableDataset
+    TimeDataset
 )
 
 
 class TestTimeDataset(unittest.TestCase):
+
+    num_workers = 0
 
     def setUp(self) -> None:
         self.adata = ad.AnnData(
@@ -124,7 +118,8 @@ class TestTimeDataset(unittest.TestCase):
 
         dl = DataLoader(
             td,
-            batch_size=5
+            batch_size=5,
+            num_workers=self.num_workers
         )
 
         self.assertEqual(len(td), 25)
@@ -153,7 +148,8 @@ class TestTimeDataset(unittest.TestCase):
 
         dl = DataLoader(
             td,
-            batch_size=10
+            batch_size=10,
+            num_workers=self.num_workers
         )
 
         self.assertEqual(len(td), 50)
@@ -396,7 +392,12 @@ class TestTimeDataset(unittest.TestCase):
 
             self.assertGreater(len(td), 0)
 
-            dl = DataLoader(td, 2, drop_last=True)
+            dl = DataLoader(
+                td,
+                2,
+                drop_last=True,
+                num_workers=self.num_workers
+            )
 
             for data in dl:
 
@@ -476,7 +477,11 @@ class TestTimeDataset(unittest.TestCase):
             return_times=True
         )
 
-        dl = DataLoader(td, batch_size=10)
+        dl = DataLoader(
+            td,
+            batch_size=10,
+            num_workers=self.num_workers
+        )
 
         t = next(iter(dl))
 
@@ -497,597 +502,3 @@ class TestTimeDatasetSparse(TestTimeDataset):
             sps.csr_matrix(X)
         )
         self.adata.obs['time'] = T
-
-
-class TestADBacked(unittest.TestCase):
-
-    dataset = None
-
-    @classmethod
-    def setUpClass(cls):
-        cls.tempdir = tempfile.TemporaryDirectory()
-        cls.filename = os.path.join(cls.tempdir.name, "tests.h5ad")
-
-    def setUp(self) -> None:
-        self.adata = ad.AnnData(
-            X
-        )
-        self.adata.obs['time'] = T
-        self.adata.layers['tst'] = self.adata.X.copy()
-        self.adata.write(self.filename)
-
-    def tearDown(self) -> None:
-        if self.dataset is not None:
-            self.dataset.close()
-
-        return super().tearDown()
-
-    def test_load_h5(self):
-
-        dataset = H5ADDataset(self.filename)
-        self.dataset = dataset
-
-        self.assertEqual(
-            len(dataset),
-            100
-        )
-
-        dataloader = DataLoader(
-            dataset,
-            batch_size=10,
-            shuffle=True
-        )
-
-        lens = [d.shape[0] for d in dataloader]
-
-        self.assertEqual(
-            len(lens),
-            10
-        )
-        self.assertListEqual(
-            lens,
-            [10] * 10
-        )
-
-    def test_h5_mask(self):
-
-        dataset = H5ADDataset(
-            self.filename,
-            obs_include_mask=np.arange(50, 100)
-        )
-        self.dataset = dataset
-
-        self.assertEqual(
-            len(dataset),
-            50
-        )
-
-        dataloader = DataLoader(
-            dataset,
-            batch_size=50,
-            shuffle=False
-        )
-
-        data = [d for d in dataloader]
-
-        self.assertEqual(
-            len(data),
-            1
-        )
-        self.assertEqual(
-            data[0].shape[0],
-            50
-        )
-
-        npt.assert_almost_equal(
-            X[np.arange(50, 100), :],
-            data[0].numpy()
-        )
-
-    def test_load_h5_layer(self):
-
-        dataset = H5ADDataset(self.filename, layer='tst')
-        self.dataset = dataset
-
-        self.assertEqual(
-            len(dataset),
-            100
-        )
-
-        dataloader = DataLoader(
-            dataset,
-            batch_size=10,
-            shuffle=True
-        )
-
-        lens = [d.shape[0] for d in dataloader]
-
-        self.assertEqual(
-            len(lens),
-            10
-        )
-        self.assertListEqual(
-            lens,
-            [10] * 10
-        )
-
-
-class TestADBackedSparse(TestADBacked):
-
-    def setUp(self) -> None:
-        self.adata = ad.AnnData(
-            sps.csr_matrix(X)
-        )
-        self.adata.obs['time'] = T
-        self.adata.layers['tst'] = self.adata.X.copy()
-        self.adata.write(self.filename)
-
-
-class TestADBackedChunk(unittest.TestCase):
-
-    dataset = None
-
-    @classmethod
-    def setUpClass(cls):
-        cls.tempdir = tempfile.TemporaryDirectory()
-        cls.filename = os.path.join(cls.tempdir.name, "tests.h5ad")
-
-    def setUp(self) -> None:
-        self.adata = ad.AnnData(
-            X
-        )
-        self.adata.obs['time'] = T
-        self.adata.layers['tst'] = self.adata.X.copy()
-        self.adata.write(self.filename)
-
-    def tearDown(self) -> None:
-        if self.dataset is not None:
-            self.dataset.close()
-
-        return super().tearDown()
-
-    def test_load_h5(self):
-
-        dataset = H5ADDatasetIterable(
-            self.filename,
-            file_chunk_size=50
-        )
-        self.dataset = dataset
-
-        self.assertEqual(
-            len(dataset.file_chunks),
-            2
-        )
-
-        npt.assert_array_equal(
-            dataset.file_chunks[0],
-            np.arange(50)
-        )
-
-        npt.assert_array_equal(
-            dataset.file_chunks[1],
-            np.arange(50, 100)
-        )
-
-        dataset.load_chunk(0)
-
-        npt.assert_array_equal(
-            dataset._data_loaded_chunk.numpy(),
-            X[0:50, :]
-        )
-
-        dataset.load_chunk(1)
-
-        npt.assert_array_equal(
-            dataset._data_loaded_chunk.numpy(),
-            X[50:100, :]
-        )
-
-        dataloader = DataLoader(
-            dataset,
-            batch_size=10
-        )
-
-        lens = [d.shape[0] for d in dataloader]
-
-        self.assertEqual(
-            len(lens),
-            10
-        )
-        self.assertListEqual(
-            lens,
-            [10] * 10
-        )
-
-    def test_h5_mask(self):
-
-        dataset = H5ADDatasetIterable(
-            self.filename,
-            obs_include_mask=np.arange(50, 100),
-            file_chunk_size=27
-        )
-        self.dataset = dataset
-
-        self.assertEqual(
-            len(dataset.file_chunks),
-            2
-        )
-
-        npt.assert_array_equal(
-            dataset.file_chunks[0],
-            np.arange(50, 77)
-        )
-
-        npt.assert_array_equal(
-            dataset.file_chunks[1],
-            np.arange(77, 100)
-        )
-
-        dataloader = DataLoader(
-            dataset,
-            batch_size=50
-        )
-
-        data = [d for d in dataloader]
-
-        self.assertEqual(
-            len(data),
-            1
-        )
-        self.assertEqual(
-            data[0].shape[0],
-            50
-        )
-
-    def test_load_h5_big_chunk(self):
-
-        dataset = H5ADDatasetIterable(
-            self.filename,
-            file_chunk_size=5000
-        )
-        self.dataset = dataset
-
-        self.assertEqual(
-            len(dataset.file_chunks),
-            1
-        )
-
-        npt.assert_array_equal(
-            dataset.file_chunks[0],
-            np.arange(100)
-        )
-
-        dataset.load_chunk(0)
-
-        npt.assert_array_equal(
-            dataset._data_loaded_chunk.numpy(),
-            X
-        )
-
-        dataloader = DataLoader(
-            dataset,
-            batch_size=10
-        )
-
-        lens = [d.shape[0] for d in dataloader]
-
-        self.assertEqual(
-            len(lens),
-            10
-        )
-        self.assertListEqual(
-            lens,
-            [10] * 10
-        )
-
-
-class TestADBackedChunkSparse(TestADBackedChunk):
-
-    def setUp(self) -> None:
-        self.adata = ad.AnnData(
-            sps.csr_matrix(X)
-        )
-        self.adata.obs['time'] = T
-        self.adata.layers['tst'] = self.adata.X.copy()
-        self.adata.write(self.filename)
-
-
-class TestADBackedStratified(unittest.TestCase):
-
-    dataset = None
-
-    @classmethod
-    def setUpClass(cls):
-        cls.tempdir = tempfile.TemporaryDirectory()
-        cls.filename = os.path.join(cls.tempdir.name, "tests.h5ad")
-
-    def setUp(self) -> None:
-        self.adata = ad.AnnData(
-            X
-        )
-        self.adata.obs['time'] = T
-
-        _strats = np.tile(["A", "B", "C"], 34)[0:100]
-        _strats[[3, 6, 10]] = 'D'
-        self.adata.obs['strat'] = _strats
-        self.adata.obs['other_strat'] = ['F'] * 49 + ['G'] * 50 + ['F']
-        self.adata.layers['tst'] = self.adata.X.copy()
-        self.adata.write(self.filename)
-
-    def tearDown(self) -> None:
-        if self.dataset is not None:
-            self.dataset.close()
-
-        return super().tearDown()
-
-    def test_load_h5(self):
-
-        dataset = H5ADDatasetStratified(
-            self.filename,
-            'strat',
-            file_chunk_size=50
-        )
-        self.dataset = dataset
-
-        self.assertEqual(
-            len(dataset.file_chunks),
-            2
-        )
-
-        pdt.assert_series_equal(
-            self.adata.obs['strat'],
-            dataset.stratification_grouping['strat']
-        )
-
-        npt.assert_array_equal(
-            dataset.file_chunks[0],
-            np.arange(50)
-        )
-
-        npt.assert_array_equal(
-            dataset.file_chunks[1],
-            np.arange(50, 100)
-        )
-
-        for c, s in enumerate([slice(0, 50), slice(50, 100)]):
-            dataset.load_chunk(c)
-
-            npt.assert_array_equal(
-                dataset._data_loaded_chunk.numpy(),
-                X[s, :]
-            )
-            self.assertEqual(
-                len(dataset._data_loaded_stratification),
-                4
-            )
-            for i in range(4):
-                _idx = np.nonzero(
-                    self.adata.obs['strat'].iloc[s].cat.codes == i
-                )[0]
-                npt.assert_array_equal(
-                    _idx,
-                    np.intersect1d(
-                        _idx,
-                        dataset._data_loaded_stratification[i]
-                    )
-                )
-
-    def test_h5_mask(self):
-
-        dataset = H5ADDatasetStratified(
-            self.filename,
-            'strat',
-            obs_include_mask=np.arange(0, 50),
-            file_chunk_size=27
-        )
-        self.dataset = dataset
-
-        self.assertEqual(
-            len(dataset.file_chunks),
-            2
-        )
-
-        npt.assert_array_equal(
-            dataset.file_chunks[0],
-            np.arange(27)
-        )
-
-        npt.assert_array_equal(
-            dataset.file_chunks[1],
-            np.arange(27, 50)
-        )
-
-        dataloader = DataLoader(
-            dataset,
-            batch_size=50
-        )
-
-        data = [d for d in dataloader]
-
-        self.assertEqual(
-            len(data),
-            1
-        )
-        self.assertEqual(
-            data[0].shape[0],
-            33
-        )
-
-    def test_dual_strat(self):
-
-        dataset = H5ADDatasetStratified(
-            self.filename,
-            ['strat', 'other_strat'],
-            discard_categories=['D'],
-            file_chunk_size=50
-        )
-
-        pdt.assert_series_equal(
-            self.adata.obs['strat'],
-            dataset.stratification_grouping['strat']
-        )
-
-        pdt.assert_series_equal(
-            self.adata.obs['other_strat'],
-            dataset.stratification_grouping['other_strat']
-        )
-
-        dataset.load_chunk(0)
-
-        npt.assert_array_equal(
-            dataset._data_loaded_chunk.numpy(),
-            X[0:50, :]
-        )
-
-        dataset.get_chunk_order(0)
-
-        self.assertEqual(
-            len(dataset._data_loaded_stratification),
-            4
-        )
-
-
-class TestADObsBacked(unittest.TestCase):
-
-    dataset = None
-    obs_dataset = None
-
-    @classmethod
-    def setUpClass(cls):
-        cls.tempdir = tempfile.TemporaryDirectory()
-        cls.filename = os.path.join(cls.tempdir.name, "tests.h5ad")
-
-        x = X.copy()
-        x[:, 0] = np.arange(x.shape[0])
-        cls.adata = ad.AnnData(
-            x
-        )
-        cls.adata.obs['time'] = T.astype(str)
-        cls.adata.obs['strat'] = np.tile(["A", "B", "C"], 34)[0:100]
-        cls.adata.write(cls.filename)
-
-    def tearDown(self) -> None:
-        if self.dataset is not None:
-            self.dataset.close()
-
-        return super().tearDown()
-
-    def test_obs_one_hot(self):
-
-        obs_dataset = H5ADDatasetObsStratified(
-            self.filename,
-            'strat',
-            file_chunk_size=27,
-            obs_columns=['time', 'strat'],
-            random_seed=876,
-            one_hot=True
-        )
-
-        self.assertEqual(
-            torch.stack([x for x in obs_dataset]).shape,
-            (99, 7)
-        )
-
-    def test_obs_classnum_discard(self):
-
-        obs_dataset = H5ADDatasetObsStratified(
-            self.filename,
-            'strat',
-            file_chunk_size=27,
-            obs_columns='strat',
-            random_seed=876,
-            one_hot=False,
-            discard_categories=['B']
-        )
-
-        self.assertEqual(
-            torch.stack([x for x in obs_dataset]).shape,
-            (66, 1)
-        )
-
-        obs_dataset.format_data(True)
-
-        self.assertEqual(
-            torch.stack([x for x in obs_dataset]).shape,
-            (66, 3)
-        )
-
-        torch.testing.assert_close(
-            torch.stack([x for x in obs_dataset]).sum(axis=0),
-            torch.Tensor([33, 0, 33])
-        )
-
-    def test_obs_classnum(self):
-
-        obs_dataset = H5ADDatasetObsStratified(
-            self.filename,
-            'strat',
-            file_chunk_size=27,
-            obs_columns='strat',
-            random_seed=876,
-            one_hot=False
-        )
-
-        self.assertEqual(
-            torch.stack([x for x in obs_dataset]).shape,
-            (99, 1)
-        )
-
-        obs_dataset.format_data(True)
-
-        self.assertEqual(
-            torch.stack([x for x in obs_dataset]).shape,
-            (99, 3)
-        )
-
-    def test_obs_alignment(self):
-
-        dataset = H5ADDatasetStratified(
-            self.filename,
-            'strat',
-            file_chunk_size=27,
-            random_seed=876
-        )
-        self.dataset = dataset
-
-        obs_dataset = H5ADDatasetObsStratified(
-            self.filename,
-            'strat',
-            file_chunk_size=27,
-            obs_columns=['time', 'strat'],
-            random_seed=876
-        )
-
-        self.assertEqual(
-            obs_dataset._data_reference.shape,
-            (100, 7)
-        )
-
-        obs_dataset._data_reference[:, 0] = torch.LongTensor(
-            np.arange(100)
-        )
-
-        for i, j in zip(DataLoader(dataset), DataLoader(obs_dataset)):
-            self.assertEqual(
-                i[0][0], j[0][0]
-            )
-
-        for data in DataLoader(
-            StackIterableDataset(
-                dataset,
-                obs_dataset
-            ),
-            batch_size=2,
-            drop_last=True
-        ):
-            self.assertEqual(
-                data[0][0][0], data[1][0][0]
-            )
-            self.assertEqual(
-                data[0].shape,
-                (2, 4)
-            )
-            self.assertEqual(
-                data[1].shape,
-                (2, 7)
-            )
